requirements.txt
requests
matplotlib
seaborn
python-dotenv
